iss-907: added test for ajc: error, props, oversvable
