+++ hystrix-contrib/hystrix-javanica/src/test/java/com/netflix/hystrix/contrib/javanica/test/spring/configuration/collapser/CollapserPropertiesTest.java
@@ -1,12 +1,12 @@
 /**
  * Copyright 2015 Netflix, Inc.
- *
+ * <p/>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p/>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p/>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,16 +15,8 @@
  */
 package com.netflix.hystrix.contrib.javanica.test.spring.configuration.collapser;
 
-import com.netflix.hystrix.HystrixEventType;
-import com.netflix.hystrix.HystrixInvokableInfo;
-import com.netflix.hystrix.HystrixRequestLog;
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixCollapser;
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixProperty;
+import com.netflix.hystrix.contrib.javanica.test.common.configuration.collapser.BasicCollapserPropertiesTest;
 import com.netflix.hystrix.contrib.javanica.test.spring.conf.AopCglibConfig;
-import com.netflix.hystrix.contrib.javanica.test.common.domain.User;
-import com.netflix.hystrix.strategy.concurrency.HystrixRequestContext;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Configurable;
@@ -32,75 +24,24 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.ExecutionException;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = {AopCglibConfig.class, CollapserPropertiesTest.CollapserPropertiesTestConfig.class})
-public class CollapserPropertiesTest {
-
+public class CollapserPropertiesTest extends BasicCollapserPropertiesTest {
 
     @Autowired
     private UserService userService;
 
-    @Test
-    public void testCollapser() throws ExecutionException, InterruptedException {
-        HystrixRequestContext context = HystrixRequestContext.initializeContext();
-        try {
-            User u1 = userService.getUser("1");
-            User u2 = userService.getUser("2");
-            User u3 = userService.getUser("3");
-            User u4 = userService.getUser("4");
-
-            assertEquals("name: 1", u1.getName());
-            assertEquals("name: 2", u2.getName());
-            assertEquals("name: 3", u3.getName());
-            assertEquals("name: 4", u4.getName());
-            assertEquals(4, HystrixRequestLog.getCurrentRequest().getAllExecutedCommands().size());
-            HystrixInvokableInfo<?> command = HystrixRequestLog.getCurrentRequest()
-                    .getAllExecutedCommands().iterator().next();
-            assertEquals("getUsers", command.getCommandKey().name());
-            // confirm that it was a COLLAPSED command execution
-            assertTrue(command.getExecutionEvents().contains(HystrixEventType.COLLAPSED));
-            // and that it was successful
-            assertTrue(command.getExecutionEvents().contains(HystrixEventType.SUCCESS));
-        } finally {
-            context.shutdown();
-        }
+    @Override
+    protected UserService createUserService() {
+        return userService;
     }
 
-    public static class UserService {
-
-        @HystrixCollapser(
-                batchMethod = "getUsers",
-                collapserKey = "GetUserCollapser", collapserProperties = {
-                @HystrixProperty(name = "maxRequestsInBatch", value = "1"),
-                @HystrixProperty(name = "timerDelayInMilliseconds", value = "200")
-        })
-        public User getUser(String id) {
-            return null;
-        }
-
-        @HystrixCommand
-        public List<User> getUsers(List<String> ids) {
-            List<User> users = new ArrayList<User>();
-            for (String id : ids) {
-                users.add(new User(id, "name: " + id));
-            }
-            return users;
-        }
-
-    }
 
     @Configurable
     public static class CollapserPropertiesTestConfig {
 
         @Bean
-        public UserService userService() {
+        public BasicCollapserPropertiesTest.UserService userService() {
             return new UserService();
         }
     }
