+++ hystrix-contrib/hystrix-javanica/src/test/java/com/netflix/hystrix/contrib/javanica/test/spring/observable/ObservableTest.java
@@ -1,12 +1,12 @@
 /**
  * Copyright 2015 Netflix, Inc.
- *
+ * <p/>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p/>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p/>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,29 +15,14 @@
  */
 package com.netflix.hystrix.contrib.javanica.test.spring.observable;
 
-
-import com.netflix.hystrix.HystrixEventType;
-import com.netflix.hystrix.HystrixRequestLog;
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
-import com.netflix.hystrix.contrib.javanica.command.ObservableResult;
+import com.netflix.hystrix.contrib.javanica.test.common.observable.BasicObservableTest;
 import com.netflix.hystrix.contrib.javanica.test.spring.conf.AopCglibConfig;
-import com.netflix.hystrix.contrib.javanica.test.common.domain.User;
-import com.netflix.hystrix.strategy.concurrency.HystrixRequestContext;
-import org.apache.commons.lang3.Validate;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Configurable;
 import org.springframework.context.annotation.Bean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import rx.Observable;
-import rx.Observer;
-import rx.functions.Action1;
-
-import static com.netflix.hystrix.contrib.javanica.test.common.CommonUtils.getHystrixCommandByKey;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 
 /**
  * Test covers "Reactive Execution" functionality.
@@ -45,104 +30,21 @@ import static org.junit.Assert.assertTrue;
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = {AopCglibConfig.class, ObservableTest.ObservableTestConfig.class})
-public class ObservableTest {
+public class ObservableTest extends BasicObservableTest {
 
     @Autowired
     private UserService userService;
 
-    @Test
-    public void testGetUserByIdObservable() {
-        HystrixRequestContext context = HystrixRequestContext.initializeContext();
-        try {
-
-            // blocking
-            assertEquals("name: 1", userService.getUser("1", "name: ").toBlocking().single().getName());
-
-            // non-blocking
-            // - this is a verbose anonymous inner-class approach and doesn't do assertions
-            Observable<User> fUser = userService.getUser("1", "name: ");
-            fUser.subscribe(new Observer<User>() {
-
-                @Override
-                public void onCompleted() {
-                    // nothing needed here
-                }
-
-                @Override
-                public void onError(Throwable e) {
-                    e.printStackTrace();
-                }
-
-                @Override
-                public void onNext(User v) {
-                    System.out.println("onNext: " + v);
-                }
-
-            });
-
-            Observable<User> fs = userService.getUser("1", "name: ");
-            fs.subscribe(new Action1<User>() {
-
-                @Override
-                public void call(User user) {
-                    assertEquals("name: 1", user.getName());
-                }
-            });
-            assertEquals(3, HystrixRequestLog.getCurrentRequest().getAllExecutedCommands().size());
-            com.netflix.hystrix.HystrixInvokableInfo getUserCommand = getHystrixCommandByKey("getUser");
-            assertTrue(getUserCommand.getExecutionEvents().contains(HystrixEventType.SUCCESS));
-        } finally {
-            context.shutdown();
-        }
-    }
-
-    @Test
-    public void testGetUserWithFallback() {
-        HystrixRequestContext context = HystrixRequestContext.initializeContext();
-        try {
-            final User exUser = new User("def", "def");
-
-            // blocking
-            assertEquals(exUser, userService.getUser(" ", "").toBlocking().single());
-            assertEquals(1, HystrixRequestLog.getCurrentRequest().getAllExecutedCommands().size());
-            com.netflix.hystrix.HystrixInvokableInfo getUserCommand = getHystrixCommandByKey("getUser");
-            // confirm that command has failed
-            assertTrue(getUserCommand.getExecutionEvents().contains(HystrixEventType.FAILURE));
-            // and that fallback was successful
-            assertTrue(getUserCommand.getExecutionEvents().contains(HystrixEventType.FALLBACK_SUCCESS));
-        } finally {
-            context.shutdown();
-        }
-    }
-
-    public static class UserService {
-
-        @HystrixCommand(fallbackMethod = "staticFallback")
-        public Observable<User> getUser(final String id, final String name) {
-            return new ObservableResult<User>() {
-                @Override
-                public User invoke() {
-                    validate(id, name);
-                    return new User(id, name + id);
-                }
-            };
-        }
-
-        private User staticFallback(String id, String name) {
-            return new User("def", "def");
-        }
-
-        private void validate(String id, String name) {
-            Validate.notBlank(id);
-            Validate.notBlank(name);
-        }
+    @Override
+    protected UserService createUserService() {
+        return userService;
     }
 
     @Configurable
     public static class ObservableTestConfig {
 
         @Bean
-        public UserService userService() {
+        public BasicObservableTest.UserService userService() {
             return new UserService();
         }
     }
